feat: complete Matrix service refactoring and improve user creation flow
## Matrix Service Refactoring
- Break down monolithic 1,960-line MatrixService into 5 specialized services:
  - MatrixClientService: authentication & client management
  - MatrixMessagingService: direct messages & communication
  - MatrixEncryptionService: E2EE management
  - MatrixSignalBridgeService: Signal integration
  - MatrixRoomService: room operations
- Maintain backward compatibility through coordinating service
- Add parseRooms() and parseCategories() methods for environment config parsing
- Preserve original service as matrix-legacy.ts backup

## User Creation & Authentication Improvements
- Fix Matrix configuration loading to check environment variables directly
- Implement automatic password reset after user creation to resolve auth issues
- Make password reset non-blocking for faster user creation response
- Fix Discourse service activation (DISCOURSE_ACTIVE=true vs True)
- Add comprehensive debug logging for email and Matrix message templates
- Increase Authentik API timeout from 15s to 30s for large user bases

## Bug Fixes
- Resolve Matrix getConfig tRPC endpoint returning null despite valid env vars
- Fix missing parseRooms/parseCategories methods after refactoring
- Enable Discourse forum post creation during user registration
- Add proper error handling and fallback behavior for Matrix operations

## Technical Improvements
- Follow Single Responsibility Principle for better maintainability
- Implement lazy initialization to prevent server startup blocking
- Add proper TypeScript interfaces and error classes
- Maintain API compatibility while enabling modular architecture
